Add in defaultFieldRenderer for path parameters component
